[Added] Volumes (Places in space of a certain radius)
3 functions:

GM:FindVolume(name, radius) -- if the name exists this function will return the info about it, otherwise it will try to find a location with the specified radius. If none is find it will return nil, otherwise a table containing radius and the pos (position).

GM:AddVolume(name, pos, radius) -- Used to manually add a location to the list

GM:RemoveVolume(name) -- Clear the lock you have on the volume so other people can use that location again